requirements.txt
asgiref==3.5.2
colorama==0.4.5
dj-database-url==1.0.0
Django==4.0.4
django-cities==0.6
django-ckeditor==6.4.2
django-heroku==0.3.1
django-js-asset==2.0.0
django-model-utils==4.2.0
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tqdm==4.64.0
tzdata==2022.1
whitenoise==6.2.0
